+++ changelog.md
@@ -1,5 +1,11 @@
 # next-graphql-react changelog
 
+## Next
+
+### Patch
+
+- Use a new internal helper function `cjsDefaultImport` to normalize the default import value from the CJS module `next/app.js` that has a `default` property, preserving the type for the various ways TypeScript may be configured.
+
 ## 15.0.1
 
 ### Patch